Corrected monte_carlo file names and CMake files
