+++ src/data.csv
@@ -1,6 +1,6 @@
-My
-name
-is
-sebas
-perilla
+My, mi
+name, nombre
+is, es
+sebas, sofia,
+perilla, vitorica
